Fixes #7 reported by @kevinrue
Previously, the 4th column of the output was the methylation level, ranging
from 0 to 1000. This is now 0-100 to match the coverage file format from
bismark. The README.md file has also been updated to actually describe this
format (not doing so was an oversight).

The version number as been bumped to 0.1.1 due to these changes.