do not recompile on "make test" and "make install"
This patch removes ".PHONY" for version.h.
".PHONY version.h" causes make to not look for the version.h file so
everything that depends on it will be rebuilt on every run of make.
This causes
    make LIBBIGWIG=/some/path/libBigWig.a
    make test
to fail as "make test" will try to recompile MethylDackel without
LIBBIGWIG set to /some/path/libBigWig.a